% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefit_cv.R
\name{benefit_cv}
\alias{benefit_cv}
\title{Cross validated c index and c for benefit}
\usage{
benefit_cv(model_fit, treat_name, folds = 10, times = 10,
  reproducible = FALSE, match_by = "benefit", data = NULL,
  replace = FALSE, matchit_args = list())
}
\arguments{
\item{model_fit}{An object as returned by model_fit or model_fit_glmnet}

\item{treat_name}{The name of the treatment variable as a string}

\item{folds}{Number of folds for k-fold cross validation}

\item{times}{Number of repetitions of the cross validation}

\item{reproducible}{Logical. If `TRUE`, then at each repetition of the cross validation and every time `matchit` is called the seed is fixed.}

\item{match_by}{One of "benefit" or "covariates" - what should the matching be based on?}

\item{data}{Must only be specified if match_by = "covariates", the data frame.}

\item{replace}{Logical, whether matching should be done with replacement. Default is `FALSE`. This argument is passed to MatchIt::matchit and overrides the parameters given in the `matchit_args` argument.}

\item{matchit_args}{A list with other arguments that are passed on to matchit. See ?MatchIt::matchit for options like distance or method of matching.}
}
\value{
An object of class cbenefit
}
\description{
Cross validated c index and c for benefit
}
\examples{
m1 <- model_fit(Surv(time, status) ~  treat + x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8,
dat_toy, model = "coxph")
benefit_cv(m1, "treat", folds = 5, times = 3)
}
