% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_plot.R
\name{calibration_plot}
\alias{calibration_plot}
\title{Calibration plots for cbenefit objects}
\usage{
calibration_plot(x, limits)
}
\arguments{
\item{x}{A cbenefit object}

\item{limits}{Optional: a vector of length two with y axis limits}
}
\value{
A list with two ggplot2 objects
}
\description{
Calibration plots for cbenefit objects
}
\examples{
m1 <- model_fit(Surv(time, status) ~  treat + x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8,
dat_toy, model = "coxph")
bf <- benefit_cv(m1, "treat", folds = 5, times = 3)
calibration_plot(bf)
}
