% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_cbenefit.R
\name{coef_cbenefit}
\alias{coef_cbenefit}
\title{Printing coefficients for c_glmnet fits}
\usage{
coef_cbenefit(mod, ...)
}
\arguments{
\item{mod}{A model as fitted by model_fit}

\item{...}{Further argumnets passed to tidy()}
}
\value{
A data frame with the estimates and confidence intervals if available
}
\description{
Printing coefficients for c_glmnet fits
}
\examples{
m1 <- model_fit(Surv(time, status) ~  treat + x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8,
dat_toy, model = "coxph")
coef_cbenefit(m1)
}
