% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\docType{data}
\name{dat_toy}
\alias{dat_toy}
\title{Toy data with treatment column and outcome}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1493 rows and 11 columns.}
\usage{
dat_toy
}
\description{
Toy data with treatment column and outcome
}
\keyword{datasets}
