% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_benefit_plot.R
\name{net_benefit_plot}
\alias{net_benefit_plot}
\title{Net benefit plots for cbenefit objects}
\usage{
net_benefit_plot(x, limits)
}
\arguments{
\item{x}{A cbenefit object}

\item{limits}{Optional: a vector of length two with y axis limits}
}
\value{
A list with two plots: one before cross validation, one after cross validation
}
\description{
Net benefit plots for cbenefit objects
}
\examples{
m2 <- model_fit(status ~  treat + x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8,
dat_toy, model = "glm")
bf <- benefit_cv(m2, "treat", folds = 5, times = 3)
net_benefit_plot(bf)
}
