% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cbenefit.R
\name{print.cbenefit}
\alias{print.cbenefit}
\title{Print cbenefit objects}
\usage{
\method{print}{cbenefit}(x, ...)
}
\arguments{
\item{x}{An object of class cbenefit}

\item{...}{Does nothing at this point}
}
\value{
Nothing
}
\description{
Print cbenefit objects
}
\examples{
m1 <- model_fit(Surv(time, status) ~  treat + x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8,
dat_toy, model = "coxph")
cb <- benefit_cv(m1, "treat", folds = 5, times = 3)
print(cb)
}
